fix: do not check invalid node_modules folder.
